requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
frozenlist==1.4.0
idna==3.4
jira==3.5.2
multidict==6.0.4
notionai-py==0.1.4
oauthlib==3.2.2
openai==0.28.0
packaging==23.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
slack-bolt==1.18.0
slack-sdk==3.22.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
